% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcMap}
\alias{calcMap}
\title{Project Y to X space}
\usage{
calcMap(x, y, row = NULL, ...)
}
\arguments{
\item{x}{numeric matrix of feature loadings,
or a \code{SingleCellExperiment}, 
or a \code{LinearEmbeddingMatrix} 
to extract them from.}

\item{y}{numeric matrix of data to project; 
e.g., features x observations expression.}

\item{row}{features to use; will internally subset 
to those present in both \code{x} and \code{y}.}
}
\value{
\code{matrix}
}
\description{
Project Y to X space
}
\examples{
spe <- readRDS(system.file("extdata", "spe.rds", package="miro"))

mtx <- replicate(26, runif(nrow(spe)))
colnames(mtx) <- letters

map <- calcMap(spe, mtx)
map[1:5, 1:5]
dim(map)

}
