% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcPca}
\alias{calcPca}
\title{Principal component analysis (PCA)}
\usage{
calcPca(x, ...)
}
\arguments{
\item{x}{numeric matrix; rows = features, columns = observations.}

\item{...}{arguments passed to \code{\link[scrapper]{runPca}}.}
}
\value{
\code{LinearEmbeddingMatrix}
}
\description{
Principal component analysis (PCA)
}
\examples{
# see ?scrapper::runPca

}
