\name{create_env}
\alias{create_env}
\title{Shortcut to create an Environment within the .GlobalEnv}
\usage{
  create_env(nm)
}
\arguments{
  \item{nm}{character string name of the environment to be
  created}
}
\description{
  This function creates an environment in the .GlobalEnv.
  Beware, it does overwrite existing environments.
}
\author{
  Matthias Bannert
}

