% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miro.R
\name{miro}
\alias{miro}
\title{spatial (transcript)omics visualization}
\arguments{
\item{dat}{...}

\item{pol, mol}{character string; polygons/molecules .parquet}

\item{pol_xy, mol_xy}{character vector; spatial coordinates}

\item{pol_id, mol_id}{character string; observation identifiers}

\item{mol_key}{character string; feature identifier}

\item{t}{character string or function; specifies whether/how to 
transform fill values; `n`one, `z`-scale, or `q`uantile scale}

\item{hl}{specifies observation to highlight;
either a logical or character vector, or a character string
specifying a logical `colData` slot or `data.frame` column;
fill of non-highlighted observations will be set to NA}

\item{na}{character string; color to use for NA values}

\item{thm}{character string; visualize using
either `"b"`lack or `"w"`hite base theme}

\item{...}{optional arguments passed between methods}
}
\value{
`ggplot`
}
\description{
spatial (transcript)omics visualization
}
\examples{
library(OSTA.data)
id <- "Xenium_HumanColon_Oliveria"
pa <- OSTA.data_load(id)
dir.create(td <- tempfile())
unzip(pa, exdir=td)

library(SpatialExperimentIO)
(spe <- readXeniumSXE(td))


}
