% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{miroBiplot}
\alias{miroBiplot}
\title{Biplot}
\usage{
miroBiplot(
  x,
  gs = 10,
  how = c("uni", "abs", "min", "max"),
  y = "PCA",
  dim = 3L,
  L = 75,
  seg = FALSE,
  rad = FALSE
)
}
\value{
\code{ggplot}
}
\description{
Biplot
}
\examples{
rds <- system.file("extdata", "spe.rds", package="miro")

spe <- readRDS(rds)
miroBiplot(spe, L=70)
miroBiplot(spe, dim=2:4)
miroBiplot(spe, gs=20, how="uni", seg=TRUE, rad=TRUE)
miroBiplot(spe, gs=grepv("^IGH", rownames(spe)), seg=TRUE)

}
