% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{miroSpatial}
\alias{miroSpatial}
\title{Spatial plot}
\usage{
miroSpatial(x, y = "PCA", dim = 3L, xy = c("x", "y"), s = 1)
}
\value{
\code{ggplot}
}
\description{
Spatial plot
}
\examples{
rds <- system.file("extdata", "spe.rds", package="miro")

spe <- readRDS(rds)
miroSpatial(spe)

}
