% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calcHex}
\alias{calcHex}
\title{Values to colors}
\usage{
calcHex(x, y = "PCA", dim = 3L, L = NULL)
}
\value{
\code{data.frame}
}
\description{
Values to colors
}
\examples{
rds <- system.file("extdata", "spe.rds", package="miro")

spe <- readRDS(rds)
df <- calcHex(spe)
head(df)

}
