% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcHex}
\alias{calcHex}
\title{Values to colors}
\usage{
calcHex(
  x,
  y = "PCA",
  dim = 3L,
  gs = NULL,
  assay = "logcounts",
  t = 0,
  L = NULL,
  llim = c(10, 90),
  alim = c(-80, 80),
  blim = alim
)
}
\arguments{
\item{x}{a \code{SingleCellExperiment} (SCE), 
or derivative thereof (e.g., SPE, SFE).}

\item{y}{character string or integer scalar
specifying which \code{reducedDim(x)} to use.}

\item{dim}{integer scalar or length-3 vector
specifying which \code{y} components to use.}

\item{gs}{features to use for projection;
only their intersection with \code{rownames(x)}
and the rotation of \code{y} will be used.}

\item{assay}{data to use for projection;
ignored when \code{gs} not specified.}

\item{t}{scalar numeric specifying rotation angle.}

\item{L}{scalar numeric in [0, 100] for luminance.}

\item{llim, alim, blim}{length-2 numeric vectors
specifying \link[colorspace]{\code{LAB}} limits;
input coordinates will be rescaled to these ranges.}
}
\value{
\code{data.frame}
}
\description{
Values to colors
}
\examples{
spe <- readRDS(system.file("extdata", "spe.rds", package="miro"))

df <- calcHex(spe)
df <- calcHex(spe, L=50)
df <- calcHex(spe, dim=c(2,4,6))

}
