% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{miroBiplot}
\alias{miroBiplot}
\title{Biplot}
\usage{
miroBiplot(x, y = "PCA", dim = 3L, gs = 5L, L = 75)
}
\value{
\code{ggplot}
}
\description{
Biplot
}
\examples{
rds <- system.file("extdata", "spe.rds", package="miro")

spe <- readRDS(rds)
miroBiplot(spe, L=70)
miroBiplot(spe, dim=2:4)
miroBiplot(spe, gs=grepv("^IGH", rownames(spe)))

}
