% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{miroBiplot}
\alias{miroBiplot}
\title{Biplot}
\usage{
miroBiplot(
  x,
  y = "PCA",
  dim = 3L,
  gs = 10,
  how = c("uni", "abs", "min", "max"),
  L = 75,
  alim = c(-100, 100),
  blim = c(-100, 100),
  seg = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{SingleCellExperiment} (SCE), 
or a derivative thereof (e.g., SPE, SFE).}

\item{y, dim, alim, blim}{parameters passed to \code{\link{calcHex}}.}

\item{gs}{scalar integer specifying how many features to select,
or a character vector specifying a fixed set of features.}

\item{how}{character string specifying how to select features:
uni = uniform sampling across color space,
abs = select based on highest absolute values,
min/max = select based on lowest/highest values.}

\item{L}{scalar numeric in [0, 100] for background luminance.}

\item{seg}{logical; whether or not to add \code{geom_segment} layer.}
}
\value{
\code{ggplot}
}
\description{
Biplot
}
\examples{
spe <- readRDS(system.file("extdata", "spe.rds", package="miro"))

miroBiplot(spe, L=100)
miroBiplot(spe, dim=2:4)
miroBiplot(spe, gs=40, how="uni")
miroBiplot(spe, gs=grepv("^IGH", rownames(spe)), seg=TRUE)

}
