% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{miroSpatial}
\alias{miroSpatial}
\title{Spatial plot}
\usage{
miroSpatial(x, xy = c("x", "y"), ...)
}
\arguments{
\item{x}{\code{SingleCellExperiment} (SCE), 
or a derivative thereof (e.g., SPE, SFE).}

\item{xy}{length-2 character vector 
specifying slots to use as spatial coordinates.
If \code{x} is a SPE, will use \code{spatialCoords}.}

\item{...}{(optional) parameters passed to 
\code{\link{calcHex}} or \code{geom_point()}.}
}
\value{
\code{ggplot}
}
\description{
Spatial plot
}
\examples{
spe <- readRDS(system.file("extdata", "spe.rds", package="miro"))

miroSpatial(spe, dim=seq(3, 5))
miroSpatial(spe, size=2, alpha=2/3)

}
